+++ package.json
@@ -50,7 +50,7 @@
     "bull": "4.4.0",
     "class-transformer": "0.5.1",
     "class-validator": "0.13.2",
-    "crypto-js": "^4.1.1",
+    "crypto-js": "4.1.1",
     "elastic-apm-node": "3.27.0",
     "firebase-admin": "10.0.2",
     "geojson": "^0.5.0",
@@ -79,7 +79,7 @@
     "@types/bull": "^3.15.7",
     "@types/chai": "^4.2.22",
     "@types/chai-as-promised": "^7.1.4",
-    "@types/crypto-js": "^4.1.1",
+    "@types/crypto-js": "4.1.1",
     "@types/express": "^4.17.13",
     "@types/geojson": "^7946.0.8",
     "@types/luxon": "^2.0.7",
