+++ src/utils/chat-crypto-service.ts
@@ -1,0 +1,28 @@
+import { ConfigService } from '@nestjs/config'
+import { Injectable } from '@nestjs/common'
+import * as CryptoJS from 'crypto-js'
+
+export interface EncryptedTextWithInitializationVector {
+  encryptedText: string
+  iv: string
+}
+
+@Injectable()
+export class ChatCryptoService {
+  key: CryptoJS.lib.WordArray
+  constructor(private configService: ConfigService) {
+    this.key = CryptoJS.enc.Utf8.parse(
+      this.configService.get('firebase.encryptionKey') ?? ''
+    )
+  }
+
+  encrypt(message: string): EncryptedTextWithInitializationVector {
+    const iv = CryptoJS.lib.WordArray.random(16)
+    const encrypted = CryptoJS.AES.encrypt(message, this.key, { iv })
+
+    return {
+      encryptedText: encrypted.ciphertext.toString(CryptoJS.enc.Base64),
+      iv: encrypted.iv.toString(CryptoJS.enc.Base64)
+    }
+  }
+}
